+++ library/eos_stp_interface.py
@@ -214,6 +214,8 @@ class EosAnsibleModule(AnsibleModule):
         keys = set(self.params).difference(self.meta_args)
         attrs = dict()
         attrs = dict([(k, self.params[k]) for k in self.params if k in keys])
+        if 'CHECKMODE' in attrs:
+            del attrs['CHECKMODE']
         return attrs
 
     def create(self):
@@ -316,7 +318,7 @@ class EosAnsibleModule(AnsibleModule):
             self.fail('Connection must define a transport')
 
         connection = pyeapi.client.make_connection(**config)
-        node = pyeapi.client.Node(connection)
+        node = pyeapi.client.Node(connection, **config)
 
         try:
             node.enable('show version')