+++ CHANGELOG.md
@@ -3,9 +3,10 @@ Arista EOS role for Ansible
 
 ## v0.2.1 (under development)
 
-- added eapi attribute to all modules
+- added connection, config attribute to all modules
 - fixes an issue with switchport trunks being set to none (issue #7)
 - support added for using enable_password over eAPI (thanks rsolomo)
+
 
 Updated Modules:
 
@@ -13,12 +14,15 @@ Updated Modules:
 * eos_vxlan
 * eos_switchport
 * eos_portchannel
-* eos_ipv4interface
+* eos_ipinterface
 
 New Modules:
 
 * eos_ethernet
 * eos_stp_interface
+* eos_vxlan_vlan
+* eos_vxlan_vtep
+* eos_ethernet
 
 
 ## v0.1.2